+++ paypaladaptive/api/__init__.py
@@ -1,0 +1,3 @@
+__author__ = 'antonagestam'
+
+from endpoints import *
