+++ requirements.txt
@@ -1,4 +1,5 @@
 Django==1.5.1
 South==0.7.6
 python-money==0.5
+pytz==2013b
 wsgiref==0.1.2
